rename package api -> client
