+++ client/application_msgraph.go
@@ -1,4 +1,4 @@
-package api
+package client
 
 import (
 	"context"
